+++ src/common/commandExecutor.ts
@@ -10,6 +10,8 @@ import {ISpawnResult} from "./node/childProcess";
 import {HostPlatform, HostPlatformId} from "./hostPlatform";
 import {ErrorHelper} from "./error/errorHelper";
 import {InternalErrorCode} from "./error/internalErrorCode";
+import * as nls from "vscode-nls";
+const localize = nls.loadMessageBundle();
 
 export enum CommandVerbosity {
     OUTPUT,
@@ -110,11 +112,11 @@ export class CommandExecutor {
                     return Q.resolve(void 0);
                 }
             }).then(() => {
-                this.logger.info("Packager stopped");
+                this.logger.info(localize("PackagerStopped", "Packager stopped"));
             });
 
         } else {
-            this.logger.warning("Packager not found");
+            this.logger.warning(localize("PackagerNotFound", "Packager not found"));
             return Q.resolve<void>(void 0);
         }
     }
@@ -217,12 +219,10 @@ export class CommandExecutor {
         switch (status) {
             case CommandStatus.Start:
                 return `Executing command: ${command}`;
-
             case CommandStatus.End:
                 return `Finished executing: ${command}`;
-
             default:
-                throw new Error("Unsupported command status");
+                throw ErrorHelper.getInternalError(InternalErrorCode.UnsupportedCommandStatus);
         }
     }
 }