+++ src/common/error/errorHelper.ts
@@ -20,6 +20,7 @@ export class ErrorHelper {
     public static wrapError(error: InternalError, innerError: Error): NestedError {
         return NestedError.getWrappedError(error, innerError);
     }
+
     public static getWarning(message: string, ...optionalArgs: any[]): InternalError {
         return new InternalError(-1, message, InternalErrorLevel.Warning);
     }
@@ -29,7 +30,7 @@ export class ErrorHelper {
     }
 
     private static getErrorMessage(errorCode: InternalErrorCode, ...optionalArgs: any[]): string {
-        return ErrorHelper.formatErrorMessage(ErrorHelper.ERROR_STRINGS[InternalErrorCode[errorCode]], ...optionalArgs);
+        return ErrorHelper.formatErrorMessage(ErrorHelper.ERROR_STRINGS[errorCode], ...optionalArgs);
     }
 
     private static formatErrorMessage(errorMessage: string, ...optionalArgs: any[]): string {