+++ src/common/error/internalError.ts
@@ -18,7 +18,7 @@ export class InternalError extends Error {
         super(message);
         this.errorCode = errorCode;
         this.errorLevel = errorLevel;
-        this.message = message;
+        this.message = errorCode > 0 ? (message + ` (error code ${this.errorCode})`) : message;
     }
 }