+++ src/common/error/internalErrorCode.ts
@@ -14,6 +14,7 @@ export enum InternalErrorCode {
         FailedToRestartPackager = 109,
         FailedToRunExponent = 110,
         FailedToPublishToExpHost = 111,
+        UnsupportedCommandStatus = 112,
 
         // Device Deployer errors
         IOSDeployNotFound = 201,
@@ -26,16 +27,21 @@ export enum InternalErrorCode {
 
         // iOS Platform errors
         IOSSimulatorNotLaunchable = 401,
+        IOSFoundMoreThanOneExecutablesCleanupBuildFolder = 402,
+        IOSCouldNotFoundExecutableInFolder = 403,
 
         // Packager errors
         OpnPackagerLocationNotFound = 501,
         OpnPackagerNotFound = 502,
         FailedToStopPackagerOnExit = 503,
+        CannotAttachToPackagerCheckPackagerRunningOnPort = 504,
+        AnotherDebuggerConnectedToPackager = 505,
 
         // React Native Project errors
         ProjectVersionNotParsable = 601,
         ProjectVersionUnsupported = 602,
         ProjectVersionNotReadable = 603,
+        NotInReactNativeFolderError = 604,
 
         // Miscellaneous errors
         TelemetryInitializationFailed = 701,
@@ -49,9 +55,17 @@ export enum InternalErrorCode {
         PlatformNotSupported = 709,
         WorkspaceNotFound = 710,
         ExpectedExponentTunnelPath = 711,
+        NotAllSuccessPatternsMatched = 712,
+        CouldNotParsePackageVersion = 713,
+        PackageNotFound = 714,
+        ReactDevtoolsIsNotInstalled = 715,
 
         // Activation errors
         CouldNotFindLocationOfNodeDebugger = 801,
+
+        // Inter Process Communication errors
+        ErrorWhileProcessingMessageInIPMSServer = 901,
+        ErrorNoPipeFound = 902,
 
         // Validating user input errors
         ExpectedIntegerValue = 1001,
@@ -60,7 +74,29 @@ export enum InternalErrorCode {
         ExpectedArrayValue = 1004,
         ExpectedObjectValue = 1005,
 
-        // Inter Process Communication errors
-        ErrorWhileProcessingMessageInIPMSServer = 901,
-        ErrorNoPipeFound = 902,
+        // Exponent errors
+        RNVersionNotSupportedByExponent = 1101,
+        UserCancelledExpoLogin = 1102,
+
+        // Android errors
+        AndroidCouldNotInstallTheAppOnAnyAvailibleDevice = 1201,
+        AndroidShellCommandTimedOut = 1202,
+        AndroidProjectNotFound = 1203,
+        AndroidMoreThanOneDeviceOrEmulator = 1204,
+        AndroidFailedToLaunchTheSpecifiedActivity = 1205,
+
+        // Windows Phone errors
+        WinRNMPPluginIsNotInstalled = 1301,
+
+        // Debugger errors
+        SourcesStoragePathIsNullOrEmpty = 1401,
+        DebuggingWontWorkReloadJSAndReconnect = 1402,
+        ReconnectionToPackagerFailedCheckForErrorsOrRestartReactNative = 1403,
+        FailedToProcessMessageFromReactNativeApp = 1404,
+        FailedToPrepareJSRuntimeEnvironment = 1405,
+        FailedToSendMessageToTheReactNativeApp = 1406,
+        ReactNativeWorkerProcessThrownAnError = 1407,
+        CouldntImportScriptAt = 1408,
+        RNMessageWithMethodExecuteApplicationScriptDoesntHaveURLProperty = 1409,
+
     }