+++ src/debugger/forkedAppWorker.ts
@@ -11,9 +11,10 @@ import { logger } from "vscode-chrome-debug-core";
 import { ErrorHelper } from "../common/error/errorHelper";
 import { IDebuggeeWorker, RNAppMessage } from "./appWorker";
 import { RemoteExtension } from "../common/remoteExtension";
+import { InternalErrorCode } from "../common/error/internalErrorCode";
 
-function printDebuggingError(message: string, reason: any) {
-    const nestedError = ErrorHelper.getNestedWarning(reason, `${message}. Debugging won't work: Try reloading the JS from inside the app, or Reconnect the VS Code debugger`);
+function printDebuggingError(error: Error, reason: any) {
+    const nestedError = ErrorHelper.getNestedError(error, InternalErrorCode.DebuggingWontWorkReloadJSAndReconnect, reason);
 
     logger.error(nestedError.message);
 }
@@ -95,7 +96,7 @@ export class ForkedAppWorker implements IDebuggeeWorker {
             this.postReplyToApp(message);
         })
         .on("error", (error: Error) => {
-            printDebuggingError("React Native worker process thrown an error", error);
+            printDebuggingError(ErrorHelper.getInternalError(InternalErrorCode.ReactNativeWorkerProcessThrownAnError), error);
         });
 
         // Resolve with port debugger server is listening on
@@ -130,14 +131,14 @@ export class ForkedAppWorker implements IDebuggeeWorker {
                             ...rnMessage,
                             url: url.format(packagerUrl),
                         };
-                        logger.verbose("Packager requested runtime to load script from " + rnMessage.url);
+                        logger.verbose(`Packager requested runtime to load script from ${rnMessage.url}`);
                         return this.scriptImporter.downloadAppScript(<string>rnMessage.url, this.projectRootPath)
                             .then((downloadedScript: DownloadedScript) => {
                                 this.bundleLoaded.resolve(void 0);
                                 return Object.assign({}, rnMessage, { url: downloadedScript.filepath });
                             });
                     } else {
-                        throw Error("RNMessage with method 'executeApplicationScript' doesn't have 'url' property");
+                        throw ErrorHelper.getInternalError(InternalErrorCode.RNMessageWithMethodExecuteApplicationScriptDoesntHaveURLProperty);
                     }
                 }
             });
@@ -147,7 +148,7 @@ export class ForkedAppWorker implements IDebuggeeWorker {
                     this.debuggeeProcess.send({ data: message });
                 }
             },
-            (reason) => printDebuggingError(`Couldn't import script at <${rnMessage.url}>`, reason));
+            (reason) => printDebuggingError(ErrorHelper.getInternalError(InternalErrorCode.CouldntImportScriptAt, rnMessage.url), reason));
 
         return promise;
     }